requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.9.1
pillow>=8.3.1
gTTS>=2.2.4
pygame>=2.0.1
numpy>=1.19.5
